requirements.txt
numpy>=1.23
Pillow>=9.0
matplotlib>=3.5
